Clustering using a bootstap node for discovery
etcd requires some external service to bootstrap the clustering. When
we have SRV support in Triton CNS we'll use that, but for now we'll
stand up a single-node cluster and bootstrap the cluster using that.

This implementation uses a shell script rather than ContainerPilot to
shim the application because we don't have a way to inject the container
IP address into the environment that we fork. This will be covered in
https://github.com/joyent/containerpilot/issues/171 and then we'll
eliminate this shell script.